+++ src/main/java/com/prgms/kokoahairshop/hairshop/controller/HairshopController.java
@@ -3,21 +3,20 @@ package com.prgms.kokoahairshop.hairshop.controller;
 import com.prgms.kokoahairshop.hairshop.dto.CreateHairshopRequest;
 import com.prgms.kokoahairshop.hairshop.dto.HairshopResponse;
 import com.prgms.kokoahairshop.hairshop.dto.ModifyHairshopRequest;
+import com.prgms.kokoahairshop.hairshop.entity.Hairshop;
 import com.prgms.kokoahairshop.hairshop.service.HairshopService;
-import java.net.URI;
-import javax.validation.Valid;
+import com.prgms.kokoahairshop.user.entity.User;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+import java.net.URI;
+
+@Slf4j
 @RestController
 @RequestMapping("/hairshops")
 public class HairshopController {
@@ -30,10 +29,13 @@ public class HairshopController {
 
     @PostMapping
     public ResponseEntity<Long> insert(
-        @Valid @RequestBody CreateHairshopRequest createHairshopRequest) {
+            @Valid @RequestBody CreateHairshopRequest createHairshopRequest, @AuthenticationPrincipal User user) {
+        if (!user.getId().equals(createHairshopRequest.getUserId())) {
+            throw new IllegalArgumentException("본인의 헤어샵만 생성할 수 있습니다.");
+        }
         HairshopResponse insert = hairshopService.insert(createHairshopRequest);
-        return ResponseEntity.created(URI.create("/api/v1/hairshops/" + insert.getId()))
-            .body(insert.getId());
+        return ResponseEntity.created(URI.create("/hairshops/" + insert.getId()))
+                .body(insert.getId());
     }
 
     @GetMapping
@@ -49,13 +51,21 @@ public class HairshopController {
 
     @PatchMapping
     public ResponseEntity<Object> modify(
-        @Valid @RequestBody ModifyHairshopRequest modifyHairshopRequest) {
+            @Valid @RequestBody ModifyHairshopRequest modifyHairshopRequest, @AuthenticationPrincipal User user) {
+        if (!user.getId().equals(modifyHairshopRequest.getUserId())) {
+            throw new IllegalArgumentException("본인의 헤어샵만 수정할 수 있습니다.");
+        }
         hairshopService.update(modifyHairshopRequest);
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deleteById(@PathVariable Long id) {
+    public ResponseEntity<Object> deleteById(@PathVariable Long id, @AuthenticationPrincipal User user) {
+        Hairshop hairshop = hairshopService.findHairshopById(id);
+        log.info("hairshop: {}", hairshop);
+        if (!user.getId().equals(hairshop.getUser().getId())) {
+            throw new IllegalArgumentException("본인의 헤어샵만 삭제할 수 있습니다.");
+        }
         hairshopService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
